api/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
email-validator
requests
openai
certifi
pyjwt[crypto]
pydantic[email]